Added base back fill build structures
The users will have to back fill data from previous builds (hundreds).
Because of this we will need to have a way to pull unfilled builds. This
will be done through a data entry methodology. At this point the table has
been created and the means to pull ALL backfill builds as well as grab
just a single random backfill build.