Added logic to display count and completion message
A few fixes have gone into place with this commit but most notably the
base structures for display on the view (and layout) have now put this on
the site for back fill.  At this point the only functionality is to
display back fill count, and gracefully handle when we get to 0 back fill
items left.