Added the basic structure of the in progress back fill
At this point there is the necessary structure in place to show *what* the
user is currently entering.  The idea here is to have an ongoing update of
the "new version info" display so the user knows the backfill data they
are entering.  This updated display should be handled entirely on the
client side.