Corrected odd behavior of post back after data add
For some reason when I post back after a successful data add, even with
the view model representing the correct data, the old values persisted.
I'm currently not sure why this happens, but in order to correct this I
check for the existence of a correct data add, and if that's the case then
just proactively clear out the input boxes from any lingering data.