Changed logic for major release determination
There was a logic issue that if there was a pre-RTM build (i.e. CTP, etc.)
then that particular version would be pulled as a major release (as it was
sorting by build number). In this case, this is not desired. In order to
get around this, we will be explicit on  how we make this determination
with an explicit data column instead of sorting logic.