Changed exception handling logic for adding a version on the database table
The logic here was originally just to give a return value of -1 for
failure and 0 for success. In this case it hid a huge bug and there was
no error logging to troubleshoot from, as it didn't throw an exception. I
need any issues to bubble up to the calling DataAccess class so that at
the very least the error can get logged.