Fixed unexplainable postback behavior for back fill add
Whenever a user would add the first back fill version, it would be
successful. But for some reason the HiddenFor elements on the view would
show client side that they were retaining the original build add, causing
a unique key violation when attempting to add the data to the underlying
version table.

The fix here was to set the Value property of the HiddenFor elements to
the explicit viewmodel value to ensure that appears correctly on the
client side for a correct postback.